Adds usability improvements.
Uses `GM.notification` to let the user know when the injector has failed.
Added a property on the global object so NovelAI's devs can deal with the brave new world of user-scripts messing with their telemetry.  I naturally have Sentry blocked, but if this gets released, other users may not.