`TrimmingService` seems to be operational.
Initial tests suggest it might be significantly slower than NovelAI's implementation.  Worst case on a story-sized input was around 50ms.  NovelAI handles all of its processing for a context in about that same amount of time.

Will have to keep an eye on it.  Perhaps inverting the approach might be the better way to go.  It's also possible that `ts-node` might be slow as heck.