Bug fixing and `trimByLength` in `TrimmingService`..
Discovered an issue where, on certain inputs, it could forget which fragment it was supposed to be trimming, and so run through the entire process again with the current sequencer, which was disastrous when it happened with `token` trimming.
Discovered an issue where the leading non-content fragments could get lost when doing a recursive trim.  This was corrected by having multiple preservation modes besides "initial".
Added a `buffer` iterable helper that cleans up a few things.
New `trimByLength` function trims according to the trimming rules, but by length instead.