WIP of end-user configuration.
God damn, does this look ugly as hell...  I dunno what I can do about the styling, considering the restylable UI NovelAI has.  I'm reluctant to just use their React modules, though that is something I'm kitted to do...
Also, testing is completely busted with this `GM_config` thing at the moment, because it just uses the DOM as soon as it loads and the DOM doesn't exist in the test environment.