I hate this library...
But it's the only one that's really still maintained, and I don't want to roll my own.  I'll just work around all its Ruby-like magic that makes it feel very inconsistent.
Also, I had to drop the idea of pulling a specific commit from the repo with Napa, since the tests could not really deal with the fact the script needed to be inlined and executed on the global scope ...or I didn't feel like figuring out how to make it work that way.
So, instead I pulled a local copy, modified it so it was a CommonJS module, adapted its wonky types, and I guess I'll deal with updating it by hand now and then.  Or not.  Who really cares...?
What?  Why not just use the `@require` directive of user-scripts like the docs suggest?  ARE YOU FUCKING INSANE!?  Yeah, let's pull code from a random URL and execute it in a privileged context!  Great idea!  The file that's there could NEVER be replaced with something malicious at some future date!
No...  Instead, I will pull a copy, vet it myself, and use that knowingly safe copy instead.  Vetting it was easy because I had to reference the code directly to figure out all its weird jank; the docs are shit and incomplete.